make routes modular and fix swagger doc for /user/create to take form data and test w/ swagger
