+++ app.js
@@ -1,12 +1,18 @@
 const express = require("express");
 const app = express();
-const swaggerJsDoc = require("swagger-jsdoc");
-const swaggerUi = require("swagger-ui-express");
-const { createProxyMiddleware } = require("http-proxy-middleware");
 const port = process.env.PORT || 5000;
+
+const { createProxyMiddleware } = require("http-proxy-middleware");
 var cors = require("cors");
 
+const userRoutes = require('./routes/userRoutes.js');
+const weatherRoutes = require('./routes/weatherRoutes');
+
 app.use(express.json());
+
+const swaggerJsDoc = require("swagger-jsdoc");
+const swaggerUi = require("swagger-ui-express");
+
 
 const swaggerOptions = {
   swaggerDefinition: {
@@ -16,17 +22,20 @@ const swaggerOptions = {
       description: "Customer API Information",
       contact: {
         name: "Deion Shallenberger",
+        email: "deion.shallenberger.170@my.csun.edu"
       },
       servers: ["http://localhost:5000"],
     },
   },
-  // ['.routes/*.js']
-  apis: ["app.js"],
+  apis: ['./routes/*.js'],
 };
+
 const swaggerDocs = swaggerJsDoc(swaggerOptions);
 
-app.use("/api-docs", swaggerUi.serve, swaggerUi.setup(swaggerDocs));
+// cors
 app.use(cors());
+
+// middleware
 app.use(
   "/api",
   createProxyMiddleware({
@@ -39,123 +48,16 @@ app.use(
   })
 );
 
-//Swagger Doc WeatherForecast API
-/**
- *  @swagger
- *  /WeatherForecastController:
- *    get:
- *      description: Use to request all weather data
- *      responses:
- *        '200':
- *          description: A successful response
- */
+// swagger docs
+app.use("/api-docs", swaggerUi.serve, swaggerUi.setup(swaggerDocs));
 
-function getRandomWeather(min, max) {
-  return Math.floor(Math.random() * (max - min) + min);
-}
-function getRandomForecast() {
-  let forecast = [
-    "Freezing",
-    "Bracing",
-    "Chilly",
-    "Cool",
-    "Mild",
-    "Warm",
-    "Balmy",
-    "Hot",
-    "Sweltering",
-    "Scorching",
-  ];
-  let random_val = Math.floor(Math.random() * forecast.length);
-  return forecast[random_val];
-}
-function getDatetime() {
-  let currentdate = new Date();
-  let datetime =
-    "Last Sync: " +
-    currentdate.getDate() +
-    "/" +
-    (currentdate.getMonth() + 1) +
-    "/" +
-    currentdate.getFullYear() +
-    " @ " +
-    currentdate.getHours() +
-    ":" +
-    currentdate.getMinutes() +
-    ":" +
-    currentdate.getSeconds();
+// weather routes
+app.use('/weather',weatherRoutes);
 
-  return datetime;
-}
-function getID() {
-  return id++;
-}
-const weather = [];
-let id = 1;
+// user routes
+app.use('/user',userRoutes);
 
-app.get("/WeatherForecastController", (req, res) => {
-  const currentWeather = {
-    id: getID(),
-    forecast: getRandomForecast(),
-    temperature: getRandomWeather(-20, 55),
-    date: getDatetime(),
-  };
-  console.log(currentWeather);
-
-  weather.push(currentWeather);
-  JSON.stringify(weather);
-  res.status(200).send(weather);
-});
-
-//Swagger Doc POST CreateUser API
-/**
- *  @swagger
- *  /CreateUser:
- *    post:
- *      description: Post User To Database
- *      parameters:
- *        - in: body
- *          name: UserName
- *          description: The user to create.
- *          schema:
- *            type: object
- *            required:
- *              - userName
- *            properties:
- *              userName:
- *                type: string
- *              password:
- *                type: string
- *      responses:
- *        '201':
- *          description: Created User
- */
-const users = [];
-
-app.post("/CreateUser", (req, res) => {
-  const user = {userName: req.body.userName, password: req.body.password }
-  users.push(user);
-  console.log(users);
-
-  res.status(201).send('Success');
-});
-
-// Extended: https://swagger.io/specification/#infoObject
-
-/**
- *  @swagger
- *  /login:
- *    post:
- *      description: Use to Login
- *      responses:
- *        '200':
- *          description: A successful response
- */
-
-app.post("/Login", (req, res) => {
-  res.status(204).send(`Username: ${firstname} Password: ${lastname}`);
-});
-
+// server port
 app.listen(port, () => {
   console.log(`Server listening on port ${port}`);
 });
